Use `GetDataReturnType` in `GridConfig` (#2797)
* Use `GetDataReturnType` in `GridConfig`

## What issue does this pull request correspond to (if any)?

Part of #2546

## What does this pull request implement?

We migrate `GridConfig#getData` to use `GetDataReturnType`. We're
following the pattern laid out in the other subclasses of `Application`.

## What changes are made in this pull request?

- Addition: Added `GridConfig.GridConfigData`.
- Breaking: The return type of `GridConfig#getData` is more precise now.

* Use `Scene.ConfiguredInstance`

Since the `Scene` can be overridden and extended by package developers,
we want to use a type that reflects what has been configured for the
`Scene`, rather than what is statically defined here in this package.